Further refactor, all hope is lost
